fixed mapping to distribution folder structure
